refactor: domain-event:store-order 모듈화에 의한 리팩터링
