+++ application/event/build.gradle
@@ -15,4 +15,10 @@ dependencies {
 
     // Swagger
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.4.0'
+
+    // domain-event
+    implementation project(':domain:domain-event:store-event')
+
+    // kafka
+    implementation project(':infra:mq:kafka-pos')
 }
