+++ application/event/src/main/java/com/event/controller/SsePosController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-import com.event.message.StoreOrderEvent;
 import com.event.service.SsePosService;
+import com.pos.event.StoreOrderEvent;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
