+++ domain/domain-event/store-event/src/main/java/com/pos/consumer/StoreOrderHandler.java
@@ -1,0 +1,10 @@
+package com.pos.consumer;
+
+import org.springframework.stereotype.Component;
+
+import com.pos.event.StoreOrderEvent;
+
+@Component
+public interface StoreOrderHandler {
+	void handleStoreOrder(String key, StoreOrderEvent storeOrderEvent);
+}
