+++ infra/mq/kafka-pos/build.gradle
@@ -5,4 +5,10 @@ jar {
     enabled = true
 }
 dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter'
+    implementation 'org.springframework.kafka:spring-kafka'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    // domain-event
+    implementation project(':domain:domain-event:store-event')
 }
