+++ settings.gradle
@@ -16,7 +16,7 @@ include(":common:base")
 
 // domain
 include(":domain:domain-pos")
-
+include(":domain:domain-event:store-event")
 // infra
 include(":infra:rdb:pos")
 include(":infra:mq:kafka-pos")
